+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- 0.3.0
+    - grid quadrants renamed to cells
+    - meta attribute storage to grid cells
+
 - 0.2.0
     - performance/cache improved for #getObjectsBetween
     - modified .position for default object property